+++ components/PushSubscribeButton/index.js
@@ -1,0 +1,1 @@
+export { PushSubscribeButton } from './PushSubscribeButton';